Save historical applicant status changes and attached comments
The current Applicant model uses an enum for saving statuses. Customer
success reached out asking to keep a list of historical changes, and
also to be able to attach comments to each and every status change. This
code is about adding that functionality.

On the backend, the ideal design is to have a new table containing each
status transition so we can (a) know what statuses an applicant had, (b)
when an applicant status was changed, and (c) why a particular status was
changed (by means of a comment column). On the frontend, we would ideally
have a dynamic rendering of the form so we can have different fields depending
on what's selected in the status input.

For this particular change, on the backend I added that new table,
`status_transitions`. Now we can access the status of an applicant by getting
it's latest associated status record. In other words, we use a `has_many` relationship
in the interest of keeping a historical track of changes. Alternatively, we could have
used a has_one but I found it hard to keep Rails from deleting the old records.

On the frontend/form, I opted for a simpler solution, relying on Rails'
`accepts_nested_attributes_for`, and rendering only the last status. One problem is that
Rails will render the `fields_for` attributes with an `id`, but in case the user is
changing the record during edit, we want it to create a new record, so I had to
double check if it was a different status and remove the `id` value. The controller has
comments with more details.

A summary of the changes:

- added Capybara so I could test the flow (it would be ideal to do it in a separate PR but I want to keep it in a single place)
- added new controller for list of status
- new StatusTransition entity, including migrations, model
- fixed previous specs
- added helper to show only the last status
- added byebug gem